requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
requests==2.28.1
python-dotenv==0.20.0
